+++ common/src/main/java/net/mehvahdjukaar/supplementaries/integration/ExposureCompat.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.supplementaries.integration;
 import io.github.mortuusars.exposure.world.item.AlbumItem;
 import io.github.mortuusars.exposure.world.item.PhotographItem;
 import io.github.mortuusars.exposure.world.item.StackedPhotographsItem;
+import io.github.mortuusars.exposure.world.item.component.album.AlbumPage;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
@@ -17,7 +18,10 @@ public class ExposureCompat {
         if (item instanceof StackedPhotographsItem spi) {
             return spi.getPhotographs(itemstack).size();
         } else if (item instanceof AlbumItem ai) {
-            return ai.getPhotographsCount(itemstack);
+            return (int) ai.getContent(itemstack).pages().stream()
+                    .map(AlbumPage::photograph)
+                    .filter(p -> p.getItem() instanceof PhotographItem)
+                    .count();
         }
         return 1;
     }